+++ project/src/components/login/login-section-locations.tsx
@@ -1,0 +1,12 @@
+function LoginSectionLocations(): JSX.Element {
+  return (
+    <section className="locations locations--login locations--current">
+      <div className="locations__item">
+        <a className="locations__item-link" href="#todo">
+          <span>Amsterdam</span>
+        </a>
+      </div>
+    </section>
+  );
+}
+export default LoginSectionLocations;
